secure-file-backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyJWT==2.8.0
pyotp==2.9.0
cryptography==42.0.2
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==11.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-health-check==3.17.0 